updated to handle multple configs

+++ source/runscript.php
@@ -7,24 +7,15 @@
 
   require_once '/usr/local/emhttp/plugins/vmbackup/include/functions.php';
 
-  // create local variables.
+  // create variables.
   // plugin name.
   $plugin = 'vmbackup';
   // user files.
   $plugin_path = '/boot/config/plugins/' . $plugin;
-  $user_script_file = $plugin_path . '/user-script.sh';
   $user_fix_snapshots_file = $plugin_path . '/user-fix-snapshots.sh';
-  $pre_script_file = $plugin_path . '/pre-script.sh';
-  $post_script_file = $plugin_path . '/post-script.sh';
   // tmp files.
   $current_datetime = date('Ymd_His');
   $tmp_plugin_path = '/tmp/vmbackup/scripts';
-  // user script tmp files.
-  $tmp_user_script_file = $tmp_plugin_path . '/user-script.sh';
-  $tmp_log_file = $tmp_plugin_path . '/'. $current_datetime .'_user-script.log';
-  $tmp_user_script_pid = $tmp_plugin_path . '/user-script.pid';
-  $tmp_pre_script_file = $tmp_plugin_path . '/pre-script.sh';
-  $tmp_post_script_file = $tmp_plugin_path . '/post-script.sh';
   // fix snapshots tmp files.
   $tmp_fix_snapshots_file = $tmp_plugin_path . '/user-fix-snapshots.sh';
   $tmp_fix_snapshots_log_file = $tmp_plugin_path . '/'. $current_datetime .'_fix-snapshots.log';
@@ -40,13 +31,49 @@
     $arg1 = "run_backup";
   }
 
+  // function to determine if a script is already running.
+  function script_running($path) {
+    // get array of pid files from the path.
+    $pid_files = glob($path);
+    // if the array is not empty, return true because a script is already running.
+    return (!empty($pid_files));
+  }
+
   if ($arg1 == "run_backup") {
+    // if no arguments were passed, set arg2 to default for backwards compatibility.
+    if (empty($arg2)) {
+      $arg2 = "default";
+    } 
+    
+    // get variables from arguments.
+    $config_name = $arg2;
+
+    // create variables.
+    // user files.
+    if (!$config_name == "default") {
+      $user_script_file = $plugin_path . '/' . $config_name . '/user-script.sh';
+      $pre_script_file = $plugin_path . '/' . $config_name . '/pre-script.sh';
+      $post_script_file = $plugin_path . '/' . $config_name . '/post-script.sh';
+    } else {
+      $user_script_file = $plugin_path . '/user-script.sh';
+      $pre_script_file = $plugin_path . '/pre-script.sh';
+      $post_script_file = $plugin_path . '/post-script.sh';
+    }
+    // tmp files.
+    $tmp_plugin_path_config = $tmp_plugin_path . '/' . $config_name;
+    // user script tmp files.
+    $tmp_user_script_file = $tmp_plugin_path_config . '/user-script.sh';
+    $tmp_log_file = $tmp_plugin_path_config . '/'. $current_datetime .'_user-script.log';
+    $tmp_user_script_pid = $tmp_plugin_path . '/' . $config_name . '.pid';
+    $tmp_pre_script_file = $tmp_plugin_path_config . '/pre-script.sh';
+    $tmp_post_script_file = $tmp_plugin_path_config . '/post-script.sh';
+
     // make directory in tmp to run script from.
-    exec("mkdir -p ".escapeshellarg($tmp_plugin_path));
+    exec("mkdir -p ".escapeshellarg($tmp_plugin_path_config));
 
     // check to see if a backup is already running.
-    if (is_file($tmp_user_script_file) || is_file($tmp_pre_script_file) || is_file($tmp_post_script_file)) {
-      file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." A backup is already running. Exiting.\n", FILE_APPEND);
+    if (script_running($tmp_plugin_path)) {
+      file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." A script is already running. Exiting.\n", FILE_APPEND);
       exit();
     }
 
@@ -57,20 +84,20 @@
     }
 
     // make sure that the user script file exists.
-    if (! is_file($user_script_file)) {
+    if (!is_file($user_script_file)) {
       // if not, exit the script.
       file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." User script file does not exist. Exiting.\n", FILE_APPEND);
       exit();
     }
 
     // remove any old logs from the tmp path.
-    $old_logs = glob($tmp_plugin_path . "/*_user-script.log");
+    $old_logs = glob($tmp_plugin_path_config . "/*_user-script.log");
     foreach ($old_logs as $log_file) {
       unlink($log_file);
     }
 
     // start logging to tmp log file.
-    file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." Starting VM Backup.\n", FILE_APPEND);
+    file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." Starting VM Backup for $config_name config.\n", FILE_APPEND);
     // log the process id of the current process running the script.
     file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." PID: ".getmypid()."\n", FILE_APPEND);
     // create text file for the process id of the current process running the script.
@@ -131,7 +158,7 @@
 
     if (is_file($tmp_pre_script_file)) {
       // build command to run pre script with logging.
-      $pre_run_cmd = $tmp_pre_script_file." >> $tmp_log_file 2>&1";
+      $pre_run_cmd = escapeshellarg($tmp_pre_script_file)." >> ".escapeshellarg($tmp_log_file)." 2>&1";
 
       // execute the command to run the pre script.
       if (file_exists($tmp_abort_now_file)) {
@@ -143,7 +170,7 @@
     }
 
     // build command to run user script with logging.
-    $run_cmd = $tmp_user_script_file." >> $tmp_log_file 2>&1";
+    $run_cmd = escapeshellarg($tmp_user_script_file)." >> ".escapeshellarg($tmp_log_file)." 2>&1";
 
     // execute the command to run the user script.
     if (file_exists($tmp_abort_now_file)) {
@@ -155,7 +182,7 @@
 
     if (is_file($tmp_post_script_file)) {
       // build command to run post script with logging.
-      $post_run_cmd = $tmp_post_script_file." >> $tmp_log_file 2>&1";
+      $post_run_cmd = escapeshellarg($tmp_post_script_file)." >> ".escapeshellarg($tmp_log_file)." 2>&1";
 
       // execute the command to run the post script.
       if (file_exists($tmp_abort_now_file)) {
@@ -196,8 +223,8 @@
     exec("mkdir -p ".escapeshellarg($tmp_plugin_path));
 
     // check to see if a backup is already running.
-    if (is_file($tmp_user_script_file) || is_file($tmp_pre_script_file) || is_file($tmp_post_script_file)) {
-      file_put_contents($tmp_fix_snapshots_log_file, date('Y-m-d H:i:s')." A backup is already running. Exiting.\n", FILE_APPEND);
+    if (script_running($tmp_plugin_path)) {
+      file_put_contents($tmp_fix_snapshots_log_file, date('Y-m-d H:i:s')." A script is already running. Exiting.\n", FILE_APPEND);
       exit();
     }
 
@@ -208,7 +235,7 @@
     }
 
     // make sure that the fix snapshots script file exists.
-    if (! is_file($user_fix_snapshots_file)) {
+    if (!is_file($user_fix_snapshots_file)) {
       // if not, exit the script.
       file_put_contents($tmp_fix_snapshots_log_file, date('Y-m-d H:i:s')." Fix Snapshots script file does not exist. Exiting.\n", FILE_APPEND);
       exit();
@@ -249,7 +276,7 @@
 
       // verify that the array is not running a parity check or rebuild. if so, exit.
       if ($parityRunning == true) {
-        file_put_contents($tmp_log_file, date('Y-m-d H:i:s')." Parity check or rebuild is in progress. Cannot run $user_script_file. Exiting.\n", FILE_APPEND);
+        file_put_contents($tmp_fix_snapshots_log_file, date('Y-m-d H:i:s')." Parity check or rebuild is in progress. Cannot run $user_fix_snapshots_file. Exiting.\n", FILE_APPEND);
         exit();
       }
     }
@@ -262,7 +289,7 @@
     file_put_contents($tmp_fix_snapshots_log_file, date('Y-m-d H:i:s')." Fix snapshots script copied to ".$tmp_fix_snapshots_file."\n", FILE_APPEND);
 
     // build command to run fix snapshots script with logging.
-    $run_cmd = $tmp_fix_snapshots_file." >> $tmp_fix_snapshots_log_file 2>&1";
+    $run_cmd = escapeshellarg($tmp_fix_snapshots_file)." >> ".escapeshellarg($tmp_fix_snapshots_log_file)." 2>&1";
 
     // execute the command to run the fix snapshots script.
     if (file_exists($tmp_abort_now_file)) {
@@ -307,52 +334,67 @@
     file_put_contents($tmp_abort_now_file, "abort");
     file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Created ".$tmp_abort_now_file.".\n", FILE_APPEND);
 
-    // check to see if both pid files don't exist.
-    if ((!is_file($tmp_user_script_pid)) && (!is_file($tmp_fix_snapshots_pid))) {
+    // check to see if a script is actually running.
+    if (!script_running($tmp_plugin_path)) {
       file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." No PID files found. Nothing to abort.", FILE_APPEND);
     }
 
-    // check for user script pid.
-    if (is_file($tmp_user_script_pid)) {
-      file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Found $tmp_user_script_pid. Attempting to kill process.\n", FILE_APPEND);
-      // attempt to kill user script process.
-      $user_script_pid = file_get_contents("$tmp_user_script_pid");
-      // try sigterm for orderly shutdown.
-      exec("pkill -SIGTERM -P " . $user_script_pid);
-      exec("killall -SIGTERM pre-script.sh");
-      exec("killall -SIGTERM user-script.sh");
-      exec("killall -SIGTERM post-script.sh");
-      // try keyboard interrupt.
-      exec("pkill -SIGINT -P " . $user_script_pid);
-      exec("killall -SIGINT pre-script.sh");
-      exec("killall -SIGINT user-script.sh");
-      exec("killall -SIGINT post-script.sh");
-      // force kill in case those didn't work.
-      exec("pkill -SIGKILL -P " . $user_script_pid);
-      exec("killall -SIGKILL pre-script.sh");
-      exec("killall -SIGKILL user-script.sh");
-      exec("killall -SIGKILL post-script.sh");
-      file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Aborted user script with pid $user_script_pid.\n", FILE_APPEND);
+    // get array of pid files from the tmp plugin path.
+    $pid_files = glob($tmp_plugin_path);
+    // go through each pid file and get the config name then try to kill that pid.
+    foreach ($pid_files as $pid_file) {
+      if (is_file($pid_file)) {
+        // get config name for pid file.
+        $config_name = pathinfo($pid_file, PATHINFO_FILENAME);
+        file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Found $pid_file. Attempting to kill process.\n", FILE_APPEND);
+        $script_pid = file_get_contents($pid_file);
+        // try sigterm for orderly shutdown.
+        exec("pkill -SIGTERM -P " . $script_pid);
+        exec("killall -SIGTERM pre-script.sh");
+        exec("killall -SIGTERM user-script.sh");
+        exec("killall -SIGTERM post-script.sh");
+        // try keyboard interrupt.
+        exec("pkill -SIGINT -P " . $script_pid);
+        exec("killall -SIGINT pre-script.sh");
+        exec("killall -SIGINT user-script.sh");
+        exec("killall -SIGINT post-script.sh");
+        // force kill in case those didn't work.
+        exec("pkill -SIGKILL -P " . $script_pid);
+        exec("killall -SIGKILL pre-script.sh");
+        exec("killall -SIGKILL user-script.sh");
+        exec("killall -SIGKILL post-script.sh");
+        file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Aborted user script with pid $script_pid for $config_name.\n", FILE_APPEND);
 
-      // remove tmp pre script file.
-      if (is_file($tmp_pre_script_file)) {
-        unlink($tmp_pre_script_file);
-        file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_pre_script_file."\n", FILE_APPEND);
+        // set variables for other files based on config name.
+        $tmp_plugin_path_config = $tmp_plugin_path . '/' . $config_name;
+        $tmp_pre_script_file = $tmp_plugin_path_config . '/pre-script.sh';
+        $tmp_user_script_file = $tmp_plugin_path_config . '/user-script.sh';
+        $tmp_post_script_file = $tmp_plugin_path_config . '/post-script.sh';
+
+        // remove tmp pre script file.
+        if (is_file($tmp_pre_script_file)) {
+          unlink($tmp_pre_script_file);
+          file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_pre_script_file."\n", FILE_APPEND);
+        }
+
+        // remove tmp user script file.
+        if (is_file($tmp_user_script_file)) {
+          unlink($tmp_user_script_file);
+          file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_user_script_file."\n", FILE_APPEND);
+        }
+
+        // remove tmp post script file.
+        if (is_file($tmp_post_script_file)) {
+          unlink($tmp_post_script_file);
+          file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_post_script_file."\n", FILE_APPEND);
+        }
+
+        // remove tmp user pid file.
+        if (is_file($tmp_user_script_pid)) {
+          unlink($tmp_user_script_pid);
+          file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_user_script_pid."\n", FILE_APPEND);
+        }
       }
-
-      // remove tmp user script file.
-      unlink($tmp_user_script_file);
-      file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_user_script_file."\n", FILE_APPEND);
-
-      // remove tmp post script file.
-      if (is_file($tmp_post_script_file)) {
-        unlink($tmp_post_script_file);
-        file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_post_script_file."\n", FILE_APPEND);
-      }
-
-      // remove tmp user pid file.
-      unlink($tmp_user_script_pid);
-      file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_user_script_pid."\n", FILE_APPEND);
     }
 
     // check for fix snapshots script pid.
@@ -361,20 +403,27 @@
       // attempt to kill fix snapshots script process.
       $fix_snapshots_pid = file_get_contents("$tmp_fix_snapshots_pid");
       // try sigterm for orderly shutdown.
+      exec("pkill -SIGTERM -P " . $fix_snapshots_pid);
       exec("killall -SIGTERM user-fix-snapshots.sh");
       // try keyboard interrupt.
+      exec("pkill -SIGINT -P " . $fix_snapshots_pid);
       exec("killall -SIGINT user-fix-snapshots.sh");
       // force kill in case those didn't work.
+      exec("pkill -SIGKILL -P " . $fix_snapshots_pid);
       exec("killall -SIGKILL user-fix-snapshots.sh");
       file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Aborted fix snapshots script with pid $fix_snapshots_pid.\n", FILE_APPEND);
 
       // remove tmp user script file.
-      unlink($tmp_fix_snapshots_file);
-      file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_fix_snapshots_file."\n", FILE_APPEND);
+      if (is_file($tmp_fix_snapshots_file)) {
+        unlink($tmp_fix_snapshots_file);
+        file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_fix_snapshots_file."\n", FILE_APPEND);
+      }
 
       // remove tmp user pid file.
-      unlink($tmp_fix_snapshots_pid);
-      file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_fix_snapshots_pid."\n", FILE_APPEND);
+      if (is_file($tmp_fix_snapshots_pid)) {
+        unlink($tmp_fix_snapshots_pid);
+        file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Removed: ".$tmp_fix_snapshots_pid."\n", FILE_APPEND);
+      }
     }
 
     // remove abort-now file.
@@ -384,6 +433,9 @@
 
     // end logging to tmp fix snapshots log file.
     file_put_contents($tmp_abort_script_log_file, date('Y-m-d H:i:s')." Finished abort script.", FILE_APPEND);
+
+    // notify the user that an abort has been finished.
+    exec('/usr/local/emhttp/plugins/dynamix/scripts/notify -s "VM Backup plugin" -d "abort finished" -i "warning" -m "$(date \'+%Y-%m-%d %H:%M\') Finished attempt to abort scripts."');
   }
 
   if ($arg1 == "show_log") {
